requirements.txt
Flask==1.1.2
Flask-Mail==0.9.1
htmltopdf==0.1.1
jwt==1.0.0
pdfkit==0.6.1
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
python-pdf==0.37
WeasyPrint==51
Werkzeug==1.0.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.4
pdfkit==0.6.1
Flask-Cors==3.0.3
gunicorn==20.0.4
PyJWT==1.7.1